test: test getData return type
fix mouse_track typo

+++ src/chaosvm/__init__.py
@@ -10,7 +10,7 @@ def prepare(
     ua="",
     href="",
     referer="",
-    mouse_track: Optional[List[Tuple[float, float]]] = None,
+    mouse_track: Optional[List[Tuple[int, int]]] = None,
 ):
     win = Window(top=True)
     if ip:
